0.8.9.904 - fix lxdm for LXDE edition - keyboard and keymap have to before encrypt hook - if you are using a non-us-keyboard and type in an encryption password, it might happen, that you can't unlock the system after reboot, because the keymap for your keyboard is not loaded - see also: http://forum.manjaro.org/index.php?topic=11170.0 - fix filesystem and partition shrink operation - add 1MB offset for ROOT size calculation - do not swapoff ZRAM - filter out empty strings from the list of swap partitions - load_html_string is deprecated since webkitgtk 1.1.1 - set_from_stock_name is deprecated since gtk 3.10 - using 'encoding=utf8' prevents an exception to rise when calling thus   using env vars in the command line (like sudo -E LANG=fr python thus.py) - update DBusException in misc - sort to_be_deleted and changelist in installation_advanced - move swap to the end of free space, and do not leave unused space - removed cairo coloring for check page - fix partitioning calls to consistently use MiB instead of mixing MiB and MB - fix ext2 mounting - mount swap after manual partitioning - update translations

+++ src/installation/auto_partition.py
@@ -35,7 +35,7 @@ import parted3.used_space as used_space
 
 """ AutoPartition class """
 
-# Partition sizes are in MB
+# Partition sizes are in MiB
 MAX_ROOT_SIZE = 30000
 
 # TODO: This higly depends on the selected DE! Must be taken into account.
@@ -57,7 +57,10 @@ def printk(enable):
 
 def unmount_all(dest_dir):
     """ Unmounts all devices that are mounted inside dest_dir """
-    subprocess.check_call(["swapoff", "-a"])
+    swaps = subprocess.check_output(["swapon", "--show=NAME", "--noheadings"]).decode().split("\n")
+    for name in filter(None, swaps):
+        if "/dev/zram" not in name:
+            subprocess.check_call(["swapoff", name])
 
     mount_result = subprocess.check_output("mount").decode().split("\n")
 
@@ -138,8 +141,14 @@ class AutoPartition(object):
         if os.path.exists("/sys/firmware/efi"):
             self.efi = True
 
+        self.separate_boot = use_luks or use_lvm or self.efi
+        logging.debug( "luks is " + str(use_luks) + ", lvm is " + str(use_lvm) \
+                       + " and efi is " + str(self.efi) \
+                       + "\ntherefore separate_boot is " + str(self.separate_boot))
+
     def mkfs(self, device, fs_type, mount_point, label_name, fs_options="", btrfs_devices=""):
         """ We have two main cases: "swap" and everything else. """
+        logging.debug("Will mkfs " + device + " as " + fs_type)
         if fs_type == "swap":
             try:
                 swap_devices = check_output("swapon -s")
@@ -163,7 +172,7 @@ class AutoPartition(object):
 
             # Make sure the fs type is one we can handle
             if fs_type not in mkfs.keys():
-                txt = _("Unkown filesystem type %s"), fs_type
+                txt = _("Unknown filesystem type %s"), fs_type
                 logging.error(txt)
                 show.error(txt)
                 return
@@ -188,7 +197,15 @@ class AutoPartition(object):
             subprocess.check_call(["mkdir", "-p", path])
 
             # Mount our new filesystem
-            subprocess.check_call(["mount", "-t", fs_type, device, path])
+
+            mopts = "rw,relatime"
+            if fs_type == "ext4":
+                mopts = "rw,relatime,data=ordered"
+            elif fs_type == "btrfs":
+                mopts = 'rw,relatime,space_cache,autodefrag,inode_cache'
+            subprocess.check_call(["mount", "-t", fs_type, "-o", mopts, device, path])
+
+            logging.debug("AutoPartition done, filesystems mounted:\n" + subprocess.check_output(["mount"]).decode())
 
             # Change permission of base directories to avoid btrfs issues
             mode = "755"
@@ -220,16 +237,26 @@ class AutoPartition(object):
         if self.efi:
             efi = self.auto_device + "2"
             boot = self.auto_device + "3"
-            swap = self.auto_device + "4"
-            root = self.auto_device + "5"
+            root = self.auto_device + "4"
+            swap = self.auto_device + "5"
             if self.home:
-                home = self.auto_device + "6"
-        else:
+                home = self.auto_device + "5"
+                swap = self.auto_device + "6"
+        elif self.luks or self.lvm:
             boot = self.auto_device + "1"
-            swap = self.auto_device + "2"
-            root = self.auto_device + "3"
+            root = self.auto_device + "2"
+            swap = self.auto_device + "3"
             if self.home:
-                home = self.auto_device + "4"
+                home = self.auto_device + "3"
+                swap = self.auto_device + "4"
+        else:
+            # self.separate_boot must be false
+            boot = ""
+            root = self.auto_device + "1"
+            swap = self.auto_device + "2"
+            if self.home:
+                home = self.auto_device + "2"
+                swap = self.auto_device + "3"
 
         if self.luks:
             if self.lvm:
@@ -247,7 +274,7 @@ class AutoPartition(object):
                     home = "/dev/mapper/cryptManjaroHome"
         elif self.lvm:
             # No LUKS but using LVM
-            lvm = swap
+            lvm = root
 
         if self.lvm:
             swap = "/dev/ManjaroVG/ManjaroSwap"
@@ -263,7 +290,8 @@ class AutoPartition(object):
         (efi_device, boot_device, swap_device, root_device, luks_devices, lvm_device, home_device) = self.get_devices()
 
         mount_devices = {}
-        mount_devices["/boot"] = boot_device
+        if self.separate_boot:
+            mount_devices["/boot"] = boot_device
         mount_devices["/"] = root_device
         mount_devices["/home"] = home_device
 
@@ -289,7 +317,8 @@ class AutoPartition(object):
 
         fs_devices = {}
 
-        fs_devices[boot_device] = "ext2"
+        if self.separate_boot:
+            fs_devices[boot_device] = "ext2"
         fs_devices[swap_device] = "swap"
 
         if self.efi:
@@ -322,8 +351,8 @@ class AutoPartition(object):
         logging.debug(_("Thus will setup LUKS on device %s"), luks_device)
 
         # Wipe LUKS header (just in case we're installing on a pre LUKS setup)
-        # For 512 bit key length the header is 2MB
-        # If in doubt, just be generous and overwrite the first 10MB or so
+        # For 512 bit key length the header is 2MiB
+        # If in doubt, just be generous and overwrite the first 10MiB or so
         subprocess.check_call(["dd", "if=/dev/zero", "of=%s" % luks_device, "bs=512", "count=20480", "status=noxfer"])
 
         if self.luks_key_pass == "":
@@ -350,7 +379,7 @@ class AutoPartition(object):
     def run(self):
         key_files = ["/tmp/.keyfile-root", "/tmp/.keyfile-home"]
 
-        # Partition sizes are expressed in MB
+        # Partition sizes are expressed in MiB
         if self.efi:
             gpt_bios_grub_part_size = 2
             efisys_part_size = 100
@@ -358,11 +387,14 @@ class AutoPartition(object):
         else:
             gpt_bios_grub_part_size = 0
             efisys_part_size = 0
-            empty_space_size = 0
+            # we start with a 1MiB offset before the first partition
+            empty_space_size = 1
 
-        boot_part_size = 200
+        boot_part_size = 0
+        if self.separate_boot:
+            boot_part_size = 200
 
-        # Get just the disk size in 1000*1000 MB
+        # Get just the disk size in MiB
         device = self.auto_device
         device_name = check_output("basename %s" % device)
         base_path = "/sys/block/%s" % device_name
@@ -415,24 +447,28 @@ class AutoPartition(object):
 
         lvm_pv_part_size = swap_part_size + root_part_size + home_part_size
 
-        logging.debug("disk_size %dMB", disk_size)
-        logging.debug("gpt_bios_grub_part_size %dMB", gpt_bios_grub_part_size)
-        logging.debug("efisys_part_size %dMB", efisys_part_size)
-        logging.debug("boot_part_size %dMB", boot_part_size)
+        logging.debug("disk_size %dMiB", disk_size)
+        logging.debug("gpt_bios_grub_part_size %dMiB", gpt_bios_grub_part_size)
+        logging.debug("efisys_part_size %dMiB", efisys_part_size)
+        logging.debug("boot_part_size %dMiB", boot_part_size)
 
         if self.lvm:
-            logging.debug("lvm_pv_part_size %dMB", lvm_pv_part_size)
+            logging.debug("lvm_pv_part_size %dMiB", lvm_pv_part_size)
 
-        logging.debug("swap_part_size %dMB", swap_part_size)
-        logging.debug("root_part_size %dMB", root_part_size)
+        logging.debug("swap_part_size %dMiB", swap_part_size)
+        logging.debug("root_part_size %dMiB", root_part_size)
 
         if self.home:
-            logging.debug("home_part_size %dMB", home_part_size)
+            logging.debug("home_part_size %dMiB", home_part_size)
 
         # Disable swap and all mounted partitions, umount / last!
         unmount_all(self.dest_dir)
 
         printk(False)
+
+        #WARNING: Our computed sizes are all in mebibytes (MiB) i.e. powers of 1024, not metric megabytes.
+        #         These are 'M' in sgdisk and 'MiB' in parted. If you use 'M' in parted you'll get MB instead of MiB,
+        #         and you're gonna have a bad time.
 
         # We assume a /dev/hdX format (or /dev/sdX)
         if self.efi:
@@ -459,17 +495,22 @@ class AutoPartition(object):
                 subprocess.check_call(['sgdisk --set-alignment="2048" --new=4:0:+%dM --typecode=4:8E00 --change-name=4:MANJARO_LVM %s'
                     % (lvm_pv_part_size, device)], shell=True)
             else:
-                subprocess.check_call(['sgdisk --set-alignment="2048" --new=4:0:+%dM --typecode=4:8200 --change-name=4:MANJARO_SWAP %s'
-                    % (swap_part_size, device)], shell=True)
-                subprocess.check_call(['sgdisk --set-alignment="2048" --new=5:0:+%dM --typecode=5:8300 --change-name=5:MANJARO_ROOT %s'
+                subprocess.check_call(['sgdisk --set-alignment="2048" --new=4:0:+%dM --typecode=4:8300 --change-name=4:MANJARO_ROOT %s'
                     % (root_part_size, device)], shell=True)
 
                 if self.home:
-                    subprocess.check_call(['sgdisk --set-alignment="2048" --new=6:0:+%dM --typecode=6:8300 --change-name=5:MANJARO_HOME %s'
+                    subprocess.check_call(['sgdisk --set-alignment="2048" --new=5:0:+%dM --typecode=5:8300 --change-name=5:MANJARO_HOME %s'
                         % (home_part_size, device)], shell=True)
+
+                    subprocess.check_call(['sgdisk --set-alignment="2048" --new=6:0:+%dM --typecode=6:8200 --change-name=6:MANJARO_SWAP %s'
+                    % (swap_part_size, device)], shell=True)
+
+                subprocess.check_call(['sgdisk --set-alignment="2048" --new=5:0:+%dM --typecode=5:8200 --change-name=5:MANJARO_SWAP %s'
+                    % (swap_part_size, device)], shell=True)
 
             logging.debug(check_output("sgdisk --print %s" % device))
         else:
+            # DOS MBR partition table
             # Start at sector 1 for 4k drive compatibility and correct alignment
             # Clean partitiontable to avoid issues!
             subprocess.check_call(["dd", "if=/dev/zero", "of=%s" % device, "bs=512", "count=2048", "status=noxfer"])
@@ -478,36 +519,48 @@ class AutoPartition(object):
             # Create DOS MBR with parted
             subprocess.check_call(["parted", "-a", "optimal", "-s", device, "mktable", "msdos"])
 
-            # Create boot partition (all sizes are in MB)
-            subprocess.check_call(["parted", "-a", "optimal", "-s", device, "mkpart", "primary", "1", str(boot_part_size)])
-            # Set boot partition as bootable
-            subprocess.check_call(["parted", "-a", "optimal", "-s", device, "set", "1", "boot", "on"])
+            if self.separate_boot:
+                # Create boot partition (all sizes are in MiB)
+                subprocess.check_call(["parted", "-a", "optimal", "-s", device, "mkpart", "primary", "1", "%dMiB" % boot_part_size])
+                # Set boot partition as bootable
+                subprocess.check_call(["parted", "-a", "optimal", "-s", device, "set", "1", "boot", "on"])
 
             if self.lvm:
                 start = boot_part_size
+                if boot_part_size is 0:
+                    start = 1
+
                 end = start + lvm_pv_part_size
                 # Create partition for lvm (will store root, swap and home (if desired) logical volumes)
-                subprocess.check_call(["parted", "-a", "optimal", "-s", device, "mkpart", "primary", str(start), "100%"])
+                subprocess.check_call(["parted", "-a", "optimal", "-s", device, "mkpart", "primary", "%dMiB" % start, "100%"])
                 # Set lvm flag
                 subprocess.check_call(["parted", "-a", "optimal", "-s", device, "set", "2", "lvm", "on"])
             else:
-                # Create swap partition
                 start = boot_part_size
-                end = start + swap_part_size
-                subprocess.check_call(["parted", "-a", "optimal", "-s", device, "mkpart", "primary", "linux-swap",
-                    str(start), str(end)])
+                if boot_part_size is 0:
+                    start = 1
 
                 # Create root partition
-                start = end
                 end = start + root_part_size
                 subprocess.check_call(["parted", "-a", "optimal", "-s", device, "mkpart", "primary",
-                    str(start), str(end)])
+                    "%dMiB" % start, "%dMiB" % end])
+
+                if not self.separate_boot:
+                    # Set this partition as bootable
+                    subprocess.check_call(["parted", "-a", "optimal", "-s", device, "set", "1", "boot", "on"])
+
 
                 if self.home:
                     # Create home partition
                     start = end
+                    end = start + home_part_size
                     subprocess.check_call(["parted", "-a", "optimal", "-s", device, "mkpart", "primary",
-                        str(start), "100%"])
+                        "%dMiB" % start, "%dMiB" % end])
+
+                # Create swap partition
+                start = end
+                subprocess.check_call(["parted", "-a", "optimal", "-s", device, "mkpart", "primary", "linux-swap",
+                    "%dMiB" % start, "100%"])
 
         printk(True)
 
@@ -539,17 +592,20 @@ class AutoPartition(object):
             subprocess.check_call(["lvcreate", "--name", "ManjaroRoot", "--size", str(int(root_part_size)), "ManjaroVG"])
 
             if not self.home:
-                # Use the remainig space for our swap volume
+                # Use the remaining space for our swap volume
                 subprocess.check_call(["lvcreate", "--name", "ManjaroSwap", "--extents", "100%FREE", "ManjaroVG"])
             else:
-                subprocess.check_call(["lvcreate", "--name", "ManjaroSwap", "--size", str(int(swap_part_size)), "ManjaroVG"])
-                # Use the remainig space for our home volume
-                subprocess.check_call(["lvcreate", "--name", "ManjaroHome", "--extents", "100%FREE", "ManjaroVG"])
+                subprocess.check_call(["lvcreate", "--name", "ManjaroHome", "--size", str(int(home_part_size)), "ManjaroVG"])
+                # Use the remaining space for our swap volume
+                subprocess.check_call(["lvcreate", "--name", "ManjaroSwap", "--extents", "100%FREE", "ManjaroVG"])
+
 
         # Make sure the "root" partition is defined first!
         self.mkfs(root_device, "ext4", "/", "ManjaroRoot")
         self.mkfs(swap_device, "swap", "", "ManjaroSwap")
-        self.mkfs(boot_device, "ext2", "/boot", "ManjaroBoot")
+        if self.separate_boot:
+            logging.debug("Boot device is " + boot_device + ", about to mkfs")
+            self.mkfs(boot_device, "ext2", "/boot", "ManjaroBoot")
 
         # Format the EFI partition
         if self.efi:
